pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "running_results"
version = "0.1.0"
description = "A framework for analyzing and visualizing running race results"
readme = { file = "README_PACKAGE.md", content-type = "text/markdown" }
authors = [{ name = "Daniel Williams" }]
license = { file = "LICENSE" }
requires-python = ">=3.7"
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "nameparser>=1.1.0",
    "requests>=2.26.0",
    "tqdm>=4.62.0",
    "typing-extensions>=4.6.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.9.0",
    "lxml>=4.6.0",
    "click>=8.0.0",
    "otter-report>=0.1.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Visualization",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
]
docs = [
    "sphinx>=4.0",
    "kentigern>=0.1.0",
]

[project.scripts]
"running-results" = "running_results.cli:cli"

[tool.setuptools.packages.find]
include = ["running_results", "running_results.*"]
